==> Risk/2.0 Pre-Processing.ipynb <==
# <center> **Home Credit Default Risk Assessment**
# <center> **Pre-Processing**

# **Introduction**

In this part of the project, I clean the data by removing inaccurate entries, infinity values, and features with just one value. I convert features with number of days to number of years. 

# **Libraries**

import pandas as pd
import numpy as np

import functions
import importlib
importlib.reload(functions)

import warnings

# **Display**

%matplotlib inline

pd.options.display.max_rows = 300000
pd.options.display.max_columns = 999
pd.options.display.max_colwidth = 500

warnings.filterwarnings("ignore")
warnings.simplefilter(action="ignore", category=FutureWarning)

pd.set_option('display.max_rows', 200)

size = 20

# **Data**

## **Load Data**

train = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\application_train.csv",
    index_col=False
)


test = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\application_test.csv",
    index_col=False
)

# **Pre-Processing**

In this section, I clean the data to get it ready for EDA and machine learning. I also change the features that use days and convert days to years. Lastly, I save the new data as a file to be used in later stages of this project.

## **Drop Features with Only One Value**

Certain column contain only one value. I drop those columns.

list_columns = functions.check_columns_with_one_uniquevalue(test)

train = train.drop(list_columns, axis = 1)

## **Remove Inaccurate Entries**

Certain columns contain inaccurate entries. I drop rows or convert the inaccurate data to NAN.

train = train[train['CODE_GENDER'] != 'XNA']
train['DAYS_EMPLOYED'][train['DAYS_EMPLOYED'] == 365243] = np.nan

## **Remove Infinity Values**

Certain columns contain rows with inf or -inf entries. I convert these entries to NAN.

train.replace([np.inf, -np.inf], np.nan, inplace=True)

## **'DAYS_BIRTH' to 'AGE' in Years**

The column DAYS_BIRTH is converted to AGE in Years.

train["DAYS_BIRTH"] = abs(train["DAYS_BIRTH"])
train["DAYS_BIRTH"] = train["DAYS_BIRTH"]/365
train.rename(columns={'DAYS_BIRTH': 'AGE'}, inplace=True)
train["AGE"].describe()  

## **'DAYS_EMPLOYED' to 'YEARS_EMPLOYED'**

The column DAYS_EMPLOYED is converted to YEARS_EMPLOYED.

train["DAYS_EMPLOYED"] = abs(train["DAYS_EMPLOYED"])
train["DAYS_EMPLOYED"] = train["DAYS_EMPLOYED"]/365
train.rename(columns={"DAYS_EMPLOYED": 'YEARS_EMPLOYED'}, inplace=True)
train["YEARS_EMPLOYED"].describe()  

## **'DAYS_ID_PUBLISH' to 'YEARS_ID_PUBLISH'**

The column DAYS_ID_PUBLISH is converted to YEARS_ID_PUBLISH.

train["DAYS_ID_PUBLISH"] = abs(train["DAYS_ID_PUBLISH"])
train["DAYS_ID_PUBLISH"] = train["DAYS_ID_PUBLISH"]/365
train.rename(columns={"DAYS_ID_PUBLISH": 'YEARS_ID_PUBLISH'}, inplace=True)
train["YEARS_ID_PUBLISH"].describe()  

## **'DAYS_REGISTRATION' to 'YEARS_REGISTRATION'**

The column DAYS_REGISTRATION is converted to YEARS_REGISTRATION.

train["DAYS_REGISTRATION"] = abs(train["DAYS_REGISTRATION"])
train["DAYS_REGISTRATION"] = train["DAYS_REGISTRATION"]/365
train.rename(columns={"DAYS_REGISTRATION": 'YEARS_REGISTRATION'}, inplace=True)
train["YEARS_REGISTRATION"].describe()  

## **'DAYS_LAST_PHONE_CHANGE' to 'YEARS_LAST_PHONE_CHANGE'**

The column DAYS_LAST_PHONE_CHANGE is converted to YEARS_LAST_PHONE_CHANGE.

train["DAYS_LAST_PHONE_CHANGE"] = abs(train["DAYS_LAST_PHONE_CHANGE"])
train["DAYS_LAST_PHONE_CHANGE"] = train["DAYS_LAST_PHONE_CHANGE"]/365
train.rename(columns={"DAYS_LAST_PHONE_CHANGE": 'YEARS_LAST_PHONE_CHANGE'}, inplace=True)
train["YEARS_LAST_PHONE_CHANGE"].describe()  

# **Save Dataframe as CSV File**

A new dataframe is created to be used in later parts of this project.

train.to_csv(r"C:\Users\Dell\Documents\AI\Risk\Data\Data\train 20.csv", index=False)

# **Summary**

> * **Data Cleaning** — I removed inaccurate entries and features with just one value. Infinity values were converted to NAN. 
> * **Days to Years** — Features that indicated days were converted to years.
> * **Cleaned Table Saved** — The clean table is saved as a CSV file for later stages of this project.
