==> Risk/5.1 Machine Learning Iterations.ipynb <==
# <center> **Home Credit Default Risk Assessment**
# <center> **Machine Learning Iterations**

# **Introduction**

In this part of the project, I utilized LightGBM's feature importance tool to identify the features with the most predictive power. With this tool, I identified 18 features that provided the same predictive power as using all the features in the application_train table.

# **Libraries**

import pandas as pd
import numpy as np

from sklearn.utils import shuffle
from sklearn.model_selection import train_test_split

import lightgbm as lgb
from sklearn.metrics import roc_auc_score

from feature_engine.encoding import WoEEncoder
from feature_engine.imputation import CategoricalImputer
from feature_engine.imputation import ArbitraryNumberImputer

import functions
import importlib
importlib.reload(functions)

import warnings

# **Display**

%matplotlib inline

pd.options.display.max_rows = 300000
pd.options.display.max_columns = 999
pd.options.display.max_colwidth = 500

warnings.filterwarnings("ignore")
warnings.simplefilter(action="ignore", category=FutureWarning)

pd.set_option('display.max_rows', 200)

# **Load Data**

train = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\Data\train 22.csv",
    index_col=False
)

## **Reduce Memory Usage**

train = functions.reduce_memory_usage(train)

## **Variables**

random_state = 101
target = 'TARGET'

## **Imputation**

ani = ArbitraryNumberImputer(arbitrary_number=-99999)
ani.fit(train)
train = ani.transform(train)

ci = CategoricalImputer(imputation_method='missing', fill_value='UNKNOWN')
ci.fit(train)
train = ci.transform(train)

## **WoE Encoder**

woe = WoEEncoder(fill_value=0.0001)
woe.fit(train, train[target])
train_encoded = woe.transform(train)

## **Train Test Split**

X = train_encoded.drop('TARGET', axis=1)
y = train_encoded['TARGET']

X, y = shuffle(X, y, random_state=random_state)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y, random_state=random_state)

### **LGBM**

model = lgb.LGBMClassifier(boosting_type='gbdt', 
                           num_leaves=31, 
                           max_depth=-1, 
                           learning_rate=0.1, 
                           n_estimators=100,
                           verbose=-1)

model.fit(X_train, y_train)

y_prob = model.predict_proba(X_test)[:, 1]  

auc_score = roc_auc_score(y_test, y_prob)
print(f"AUC Score: {auc_score:.2f}")

## **Feature Importance**

feature_importance = model.feature_importances_
feature_names = X.columns

importance_df = pd.DataFrame({'Feature': feature_names, 'Importance': feature_importance})
importance_df = importance_df.sort_values(by='Importance', ascending=False).reset_index(drop=True)

importance_df

## **Keep Columns**

columns = ['SK_ID_CURR',
                   'ANNUITY_TO_CREDIT_RATIO',
                   'EXT_SOURCE_3',
                   'EXT_SOURCE_1',
                   'EXT_SOURCE_2',
                   'EXT_SOURCE_MEAN',
                   'ANNUAL_PAYMENT_TO_CREDIT_RATIO', 
                   'AGE', 
                   'YEARS_ID_PUBLISH',
                   'AMT_ANNUITY',
                   'AMT_GOODS_PRICE',
                   'ANNUITY_TO_INCOME_RATIO',
                   'ORGANIZATION_TYPE',
                   'AMT_CREDIT',
                   'YEARS_EMPLOYED_RATIO',
                   'YEARS_REGISTRATION',
                   'YEARS_LAST_PHONE_CHANGE',
                   'YEARS_EMPLOYED_AGE_PRODUCT',
                   'INCOME_TO_AGE_RATIO',
                   'CREDIT_TO_AGE_RATIO',
                   'EXT_SOURCE_PRODUCT',
                   'DEBT_TO_INCOME_RATIO',
                   'REGION_POPULATION_RELATIVE',
                   'OWN_CAR_AGE',
                   'CODE_GENDER',
                   'CREDIT_PER_DEPENDENT',
                   'TARGET'] 
train = train[columns]

train.shape

### **LGBM**

X = train.drop('TARGET', axis=1)
y = train['TARGET']

X, y = shuffle(X, y, random_state=random_state)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y, random_state=random_state)

model = lgb.LGBMClassifier(boosting_type='gbdt', 
                           num_leaves=31, 
                           max_depth=-1, 
                           learning_rate=0.1, 
                           n_estimators=100,
                           verbose=-1)

model.fit(X_train, y_train)

y_prob = model.predict_proba(X_test)[:, 1]  

auc_score = roc_auc_score(y_test, y_prob)
print(f"AUC Score: {auc_score:.2f}")

## **Feature Importance**

feature_importance = model.feature_importances_
feature_names = X.columns

importance_df = pd.DataFrame({'Feature': feature_names, 'Importance': feature_importance})
importance_df = importance_df.sort_values(by='Importance', ascending=False).reset_index(drop=True)

importance_df

## **Drop Columns**

columns = ['CODE_GENDER',
           'OWN_CAR_AGE',
           'AMT_CREDIT',
           'YEARS_EMPLOYED_RATIO',
           'CREDIT_TO_AGE_RATIO',
           'DEBT_TO_INCOME_RATIO',
           'CREDIT_PER_DEPENDENT',
           'EXT_SOURCE_PRODUCT'
                ] 
train = train.drop(columns=columns)

train.shape

### **LGBM**

X = train.drop('TARGET', axis=1)
y = train['TARGET']

X, y = shuffle(X, y, random_state=random_state)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y, random_state=random_state)

model = lgb.LGBMClassifier(boosting_type='gbdt', 
                           num_leaves=31, 
                           max_depth=-1, 
                           learning_rate=0.1, 
                           n_estimators=100,
                           verbose=-1)

model.fit(X_train, y_train)

y_prob = model.predict_proba(X_test)[:, 1]  

auc_score = roc_auc_score(y_test, y_prob)
print(f"AUC Score: {auc_score:.2f}")

## **Feature Importance**

feature_importance = model.feature_importances_
feature_names = X.columns

importance_df = pd.DataFrame({'Feature': feature_names, 'Importance': feature_importance})
importance_df = importance_df.sort_values(by='Importance', ascending=False).reset_index(drop=True)

importance_df

# **Save Dataframe as CSV File**

train.to_csv(r"C:\Users\Dell\Documents\AI\Risk\Data\Data\train 23.csv", index=False)

# **Summary**

> * **Iterative Machine Learning** — I used LightGBM to iteratively reduce the number of features utilized by the model without sacrificing performance. 
> * **Reduction of Number of Features** — From 122 features, I extracted 18 without any loss in performance.
