==> Risk/7.0 Previous.ipynb <==
# <center> **Home Credit Default Risk Assessment**
# <center> **Previous Applications Dataset**

# **Introduction**

In this part of the project, I aggregate features in the previous_payments table that I will merge with the main application_train table at a later stage. The new aggregated features will prove to have good predictive abilities.

# **Libraries**

import pandas as pd
import numpy as np

from feature_engine.imputation import CategoricalImputer
from feature_engine.imputation import ArbitraryNumberImputer

import functions
import importlib
importlib.reload(functions)

import warnings

# **Display**

%matplotlib inline

pd.options.display.max_rows = 300000
pd.options.display.max_columns = 999
pd.options.display.max_colwidth = 500

warnings.filterwarnings("ignore")
warnings.simplefilter(action="ignore", category=FutureWarning)

pd.set_option('display.max_rows', 200)

size = 20

# **Data**

## **Load Data**

previous = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\previous_application.csv",
    index_col=False
)

## **Reduce Memory Usage**

Changing datatypes to a lower level to save on system resources.

previous = functions.reduce_memory_usage(previous)

## **Clean Bad Data**

Here I replace inaccurate data entries with NAN.

previous['DAYS_FIRST_DRAWING'][previous['DAYS_FIRST_DRAWING'] == 365243.0] = np.nan
previous['DAYS_FIRST_DUE'][previous['DAYS_FIRST_DUE'] == 365243.0] = np.nan
previous['DAYS_LAST_DUE_1ST_VERSION'][previous['DAYS_LAST_DUE_1ST_VERSION'] == 365243.0] = np.nan
previous['DAYS_LAST_DUE'][previous['DAYS_LAST_DUE'] == 365243.0] = np.nan
previous['DAYS_TERMINATION'][previous['DAYS_TERMINATION'] == 365243.0] = np.nan

## **Remove Infinity Values**

Here I replace inf and -inf with NAN.

previous.replace([np.inf, -np.inf], np.nan, inplace=True)

## **Missing Values**

functions.MissingValues(previous)

## **Drop Features with More than 50% of Missing Values**

Here I drop features that have more than 50% missing values.

columns = ['RATE_INTEREST_PRIMARY', 'RATE_INTEREST_PRIVILEGED', 'RATE_DOWN_PAYMENT', 'AMT_DOWN_PAYMENT']
previous = previous.drop(columns, axis=1)

## **Imputation**

Imputation of missing values. Numerical missing values were imputed with an arbitrary number. Categorical missing values were imputed by "UNKNOWN". This is a deliberate choice to not introduce new patterns in the data.

ani = ArbitraryNumberImputer(arbitrary_number=-99999)
ani.fit(previous)
previous = ani.transform(previous)

ci = CategoricalImputer(imputation_method='missing', fill_value='UNKNOWN')
ci.fit(previous)
previous = ci.transform(previous)

## **Aggregation**

Aggregation and creation of new features.

previous = previous.groupby('SK_ID_CURR').agg({
    'SK_ID_PREV': 'count',  
    'AMT_ANNUITY': 'mean',    
    'DAYS_DECISION': ['mean', 'max', 'min'],  
    'CNT_PAYMENT': ['sum'],
    'DAYS_FIRST_DUE': lambda x: x.max() - x.min(),
    'DAYS_LAST_DUE': lambda x: x.max() - x.min()
}).reset_index()

previous.columns = ['_'.join(col).strip() if type(col) is tuple else col for col in previous.columns]

previous = previous.rename(columns={
    'SK_ID_CURR_': 'SK_ID_CURR', 
    'SK_ID_PREV_count': 'NUM_PREVIOUS_APPLICATIONS',
    'AMT_ANNUITY_mean': 'AVG_ANNUITY_AMOUNT',
    'DAYS_DECISION_mean': 'AVG_DAYS_DECISION',
    'DAYS_DECISION_max': 'MAX_DAYS_DECISION',
    'DAYS_DECISION_min': 'MIN_DAYS_DECISION',
    'CNT_PAYMENT_sum': 'SUM_CNT_PAYMENT',
    'DAYS_FIRST_DUE_<lambda>': 'RANGE_DAYS_FIRST_DUE',
    'DAYS_LAST_DUE_<lambda>': 'RANGE_DAYS_LAST_DUE',
})

# **Save Dataframe as CSV File**

A new dataframe is created to be used in later parts of this project.

previous.to_csv(r"C:\Users\Dell\Documents\AI\Risk\Data\Data\previous 25.csv", index=False)

# **Summary**

> * **Data Cleaning** — I cleaned this table of missing values, bad data and infinity values.
> * **previous_applications Table** — I used domain knowledge gained from research to aggregate and create new features from those in the previous_applications table. 
> * **Merge** — I will later merge this new table with the main application_train table in Notebook 10.0.
