==> Risk/10.0 Final Table.ipynb <==
# <center> **Home Credit Default Risk Assessment**
# <center> **Creation of Final Table**

# **Introduction**

In this part of the project, I merge the four tables created in earlier parts of this project into one table for final modeling.

# **Libraries**

import pandas as pd
import numpy as np

import functions
import importlib
importlib.reload(functions)

import warnings

# **Display**

%matplotlib inline

pd.options.display.max_rows = 300000
pd.options.display.max_columns = 999
pd.options.display.max_colwidth = 500

warnings.filterwarnings("ignore")
warnings.simplefilter(action="ignore", category=FutureWarning)

pd.set_option('display.max_rows', 200)

size = 20

# **Load Data**

train = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\Data\train 23.csv",
    index_col=False)

bureau = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\Data\bureau 24.csv",
    index_col=False)

previous = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\Data\previous 25.csv",
    index_col=False)

install = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Risk\Data\Data\install 26.csv",
    index_col=False)

## **Reduce Memory Usage**

Changing datatypes to a lower level to save on system resources.

train = functions.reduce_memory_usage(train)

bureau = functions.reduce_memory_usage(bureau)

previous = functions.reduce_memory_usage(previous)

install = functions.reduce_memory_usage(install)

# **Merge Dataframes**

Merging all the created table to create one table that will be used for machine learning in the next stage of the project.

data = train.merge(bureau, on='SK_ID_CURR').merge(previous, on='SK_ID_CURR').merge(install, on='SK_ID_CURR')

data.shape

# **Save Dataframe as CSV File**

A new dataframe is created to be used in later parts of this project.

data.to_csv(r"C:\Users\Dell\Documents\AI\Risk\Data\Data\data 27.csv", index=False)

# **Summary**

> * **train** — This table was created by selecting a subset of features and creating new features in Notebook 5.0. 
> * **bureau** — This table was created by aggregation and feature creation in the bureau and burea_balance tables in Notebook 6.0.  
> * **previous** — This table was created by aggregation and feature creation in in Notebook 7.0.
> * **install** — This table was created by aggregation and feature creation in Notebook 8.0.
> * **Final Table** — The final table was created in this notebook by merging the above 4 tables.
