==> Risk/12.0 Conclusions.ipynb <==
# <center> **Home Credit Default Risk Assessment**
# <center> **Conclusions**
# <center> **Suggestions for Improvement**

# **Conclusions**

Here is a summary of the conclusions that may be drawn from this report. 

>* **The Analysis of the Data:** I reviewed approximately 50,000,000 datapoints in 5 tables related to clients in the Home Credit Default Risk Assessment dataset. <br> 
>* **The Goal of the Project:** The goal of this project was to find a model that could predict if a client will default on a loan. I further wanted to make this determination with as few feature and in the most transparent manner possible.<br>
>* **Missing Data:** There was a great deal of missing data in many of the features. I imputed this missing data by assigning a random number to the numerical columns and labeling missing categorical data with the text, "UNKNOWN"."  <br>
>* **Models:** I utilized the following models: Logistic Regression, Random Forest, Light Gradient Boosting Machine (LGBM).
>* **Logistic Regression:** ROC-AUC Score: 0.63.  <br>
>* **Random Forest:** ROC-AUC Score: 0.73.   <br>
>* **XGBClassifier:** ROC-AUC Score: 0.72.   <br>
>* **Light Gradient Boosting (LGBM):** OC-AUC Score: 0.74.   <br>
>* **OPTUNA:** I used OPTUNA to tune the hyperparameters of the LGBM model. <br>
>* **Final Model (LGBM) with Hyperparameter Tuning:** ROC-AUC Score: 0.77.<br>

# **Suggestions for Improvement**

This report has certain weaknesses. In this section, I outline those weaknesses and indicate some remedies. 

>* **Domain Knowledge:** It is best if the data scientist has adequate domain knowledge on the topic of the analysis. I do not have any expertise in the financial sector. By doing some research, I tried to augment this lack. However, there may be parts of the data that I have overlooked that may have been important and I may have given importance to parts that may have had little significance. <br>
>* **More Complete Data:** As mentioned earlier, there was missing data in many of the features. Less missing data could have improved the performance of the models.<br>  
>* **Pandas:** Continue to learn to utilize more optimized Pandas techniques and algorithms.<br>
>* **Seaborn and Matplotlib:** Continue to improve my knowledge of Seaborn and Matplotlib for creating visualizations. <br>
>* **Machine Learning:** Continue to improve on my capabilities with various models. <br>
>* **Python Code:** Continue to write better and more efficient Python code. <br>
>* **Clean Code:** Continue to adhere to the principles of writing clean code. <br>
>* **Readability and Efficiency:** Continue to improve my skills to find the delicate balance between readability and efficiency in coding.<br>
