==> Risk/13.0 Deployment.ipynb <==
# <center> **Home Credit Default Risk Assessment**
# <center> **Deployment**

## **Default Prediction Web Application**

This deployment uses 3 features to demonstrate how the web application works. The rest of the features are set as constants. This application can be scaled to include additional features. Click on the link below.

# <center> [Default Prediction](https://k7wwqra2pgijjzzhg3pakz.streamlit.app/)
